article_six_c/requirements.txt
openai
pyyaml
python-dotenv
jinja2
numpy
tenacity
pydantic
lmstudio
os
pathlib
datetime
